Moved api Models to com.monumental.controllers.helpers
