+++ src/main/java/com/monumental/controllers/MonumentController.java
@@ -4,8 +4,8 @@ import com.monumental.exceptions.ResourceNotFoundException;
 import com.monumental.models.Image;
 import com.monumental.models.Monument;
 import com.monumental.models.Reference;
-import com.monumental.models.api.CreateMonumentRequest;
-import com.monumental.models.api.MonumentAboutPageStatistics;
+import com.monumental.controllers.helpers.CreateMonumentRequest;
+import com.monumental.controllers.helpers.MonumentAboutPageStatistics;
 import com.monumental.repositories.MonumentRepository;
 import com.monumental.services.MonumentService;
 import com.monumental.util.csvparsing.BulkCreateResult;
