+++ src/test/java/com/monumental/services/integrationtest/MonumentServiceIntegrationTests.java
@@ -2,7 +2,7 @@ package com.monumental.services.integrationtest;
 
 import com.monumental.models.Monument;
 import com.monumental.models.Tag;
-import com.monumental.models.api.MonumentAboutPageStatistics;
+import com.monumental.controllers.helpers.MonumentAboutPageStatistics;
 import com.monumental.repositories.MonumentRepository;
 import com.monumental.repositories.TagRepository;
 import com.monumental.services.MonumentService;
